+++ b4n/ui/presentation/content/select.rs
@@ -1,15 +1,40 @@
 use b4n_config::keys::KeyCombination;
 use b4n_tui::{MouseEvent, MouseEventKind, TuiEvent};
 use crossterm::event::{KeyCode, KeyModifiers};
-use ratatui::{layout::Rect, widgets::Widget};
+use ratatui::{layout::Rect, style::Color, widgets::Widget};
 
-use crate::ui::presentation::{Content, content::search::PagePosition};
+use crate::ui::presentation::{Content, content::search::ContentPosition};
+
+/// Represents a text selection defined by a `start` and `end` position in the content.
+#[derive(Clone)]
+pub struct Selection {
+    pub start: ContentPosition,
+    pub end: ContentPosition,
+}
+
+impl Selection {
+    /// Creates new [`Selection`] instance.
+    pub fn new(start: ContentPosition, end: ContentPosition) -> Self {
+        Self { start, end }
+    }
+
+    /// Returns `true` if `end` comes after `start` in document order.
+    pub fn is_end_after_start(&self) -> bool {
+        is_sorted(self.start, self.end)
+    }
+
+    /// Returns the two positions sorted in document order, ensuring that the first is always the earlier position.
+    pub fn sorted(&self) -> (ContentPosition, ContentPosition) {
+        sort(self.start, self.end)
+    }
+}
 
 /// Context for the selected text.
 #[derive(Default)]
 pub struct SelectContext {
-    pub start: Option<PagePosition>,
-    pub end: Option<PagePosition>,
+    pub start: Option<ContentPosition>,
+    pub end: Option<ContentPosition>,
+    init: Option<ContentPosition>,
 }
 
 impl SelectContext {
@@ -17,14 +42,27 @@ impl SelectContext {
     pub fn clear_selection(&mut self) {
         self.start = None;
         self.end = None;
+        self.init = None;
     }
 
-    /// Returns selection range if anything is selected.
-    pub fn get_selection(&self) -> Option<(PagePosition, PagePosition)> {
-        let (Some(start), Some(end)) = (self.start, self.end) else {
-            return None;
-        };
-        Some(sort(start, end))
+    /// Clears the current selection start (if end is not set) or adjusts the `init` of the selection
+    /// to match situation when cursor is after selection start.
+    pub fn adjust_selection(&mut self) {
+        if self.end.is_none() {
+            self.start = None;
+            self.init = None;
+        } else {
+            self.init = self.start;
+        }
+    }
+
+    /// Returns a text selection.
+    pub fn get_selection(&self) -> Option<Selection> {
+        if let (Some(start), Some(end)) = (self.start, self.end) {
+            Some(Selection::new(start, end))
+        } else {
+            None
+        }
     }
 
     /// Process UI key/mouse event.
@@ -32,8 +70,8 @@ impl SelectContext {
         &mut self,
         event: &TuiEvent,
         content: &mut T,
-        page_start: &mut PagePosition,
-        cursor: Option<PagePosition>,
+        page_start: &mut ContentPosition,
+        cursor: Option<ContentPosition>,
         area: Rect,
     ) {
         match event {
@@ -44,23 +82,8 @@ impl SelectContext {
 
     /// Updates selection end to the current cursor position only for appropriate key combinations.\
     /// **Note** that it must be executed only in edit mode and after processing edit events.
-    pub fn process_event_final(&mut self, event: &TuiEvent, cursor: PagePosition) {
+    pub fn process_event_final<T: Content>(&mut self, event: &TuiEvent, content: &T, cursor: ContentPosition) {
         let TuiEvent::Key(key) = event else {
-            return;
-        };
-
-        if key.modifiers != KeyModifiers::SHIFT {
-            return;
-        }
-
-        if is_allowed_key_code(key.code) {
-            self.end = Some(cursor);
-        }
-    }
-
-    fn process_key_event(&mut self, key: &KeyCombination, cursor: Option<PagePosition>) {
-        let Some(cursor) = cursor else {
-            // if we are not in edit mode just return
             return;
         };
 
@@ -69,8 +92,36 @@ impl SelectContext {
             return;
         }
 
+        if let Some(init) = self.init
+            && is_allowed_key_code(key.code)
+        {
+            if is_sorted(init, cursor) {
+                self.start = self.init;
+                if init == cursor {
+                    self.end = None;
+                } else {
+                    self.end = Some(decrement_cursor_x(cursor, content));
+                }
+            } else {
+                self.start = Some(decrement_cursor_x(init, content));
+                self.end = Some(cursor);
+            }
+        }
+    }
+
+    fn process_key_event(&mut self, key: &KeyCombination, cursor: Option<ContentPosition>) {
+        let Some(cursor) = cursor else {
+            // if we are not in the edit mode just return
+            return;
+        };
+
+        if key.modifiers != KeyModifiers::SHIFT {
+            return;
+        }
+
         if is_allowed_key_code(key.code) {
-            if self.start.is_none() {
+            if self.init.is_none() {
+                self.init = Some(cursor);
                 self.start = Some(cursor);
             }
         } else {
@@ -82,40 +133,50 @@ impl SelectContext {
         &mut self,
         mouse: &MouseEvent,
         content: &mut T,
-        page_start: &mut PagePosition,
+        page_start: &mut ContentPosition,
         area: Rect,
     ) {
         if mouse.kind == MouseEventKind::LeftDoubleClick {
-            let pos = get_position_in_content(area, content, *page_start, mouse.column, mouse.row);
-            if let Some((start, end)) = content.word_bounds(pos.y, pos.x) {
-                self.start = Some(PagePosition { x: start, y: pos.y });
-                self.end = Some(PagePosition { x: end, y: pos.y });
+            if let Some(pos) = get_position_in_content(area, content, *page_start, None, mouse.column, mouse.row)
+                && let Some((start, end)) = content.word_bounds(pos.y, pos.x)
+            {
+                self.init = Some(ContentPosition { x: start, y: pos.y });
+                self.start = self.init;
+                self.end = Some(ContentPosition { x: end, y: pos.y });
             }
         } else if mouse.kind == MouseEventKind::LeftClick {
-            let pos = get_position_in_content(area, content, *page_start, mouse.column, mouse.row);
-            self.start = Some(pos);
+            self.init = get_position_in_content(area, content, *page_start, None, mouse.column, mouse.row);
+            self.start = self.init;
             self.end = None;
         } else if mouse.kind == MouseEventKind::LeftDrag {
             scroll_page_if_needed(area, page_start, content, mouse.column, mouse.row);
-            let pos = get_position_in_content(area, content, *page_start, mouse.column, mouse.row);
-            self.end = Some(pos);
+            self.end = get_position_in_content(area, content, *page_start, self.init, mouse.column, mouse.row);
+            if let Some(init) = self.init
+                && let Some(end) = self.end
+            {
+                if is_sorted(init, end) {
+                    self.start = self.init;
+                } else {
+                    self.start = Some(decrement_cursor_x(init, content));
+                }
+            }
         }
     }
 }
 
-fn scroll_page_if_needed<T: Content>(area: Rect, page_start: &mut PagePosition, content: &mut T, mouse_x: u16, mouse_y: u16) {
+fn scroll_page_if_needed<T: Content>(area: Rect, page_start: &mut ContentPosition, content: &mut T, mouse_x: u16, mouse_y: u16) {
     // scroll page vertically while dragging
     if mouse_y > (area.y + area.height).saturating_sub(3) {
         page_start.y += 2;
     } else if mouse_y < area.y + 3 {
-        page_start.y = page_start.y.saturating_sub(2)
+        page_start.y = page_start.y.saturating_sub(2);
     }
 
     // scroll page horizontally while dragging
     if mouse_x > (area.x + area.width).saturating_sub(3) {
         page_start.x += 2;
     } else if mouse_x < area.x + 3 {
-        page_start.x = page_start.x.saturating_sub(2)
+        page_start.x = page_start.x.saturating_sub(2);
     }
 
     // apply page start constraints
@@ -131,31 +192,71 @@ fn scroll_page_if_needed<T: Content>(area: Rect, page_start: &mut PagePosition, 
 fn get_position_in_content<T: Content>(
     area: Rect,
     content: &T,
-    page_start: PagePosition,
+    page_start: ContentPosition,
+    selection_start: Option<ContentPosition>,
     screen_x: u16,
     screen_y: u16,
-) -> PagePosition {
+) -> Option<ContentPosition> {
     let x = page_start.x.saturating_add(screen_x.saturating_sub(area.x).into());
     let y = page_start.y.saturating_add(screen_y.saturating_sub(area.y).into());
 
-    let x = x.min(content.line_size(y));
-    PagePosition { x, y }
+    if y >= content.len() {
+        let y = content.len().saturating_sub(1);
+        let x = content.line_size(y).saturating_sub(1);
+        return Some(ContentPosition { x, y });
+    }
+
+    let line_len = content.line_size(y);
+    if let Some(start) = selection_start {
+        // we already have a selection start
+        if start.y == y && start.x >= line_len && x >= line_len {
+            // selection started on the same line and outside the text, return nothing
+            None
+        } else if is_sorted(ContentPosition { x, y }, start) {
+            // selection end is before selection start
+            Some(ContentPosition { x: x.min(line_len), y })
+        } else {
+            let x = x.min(line_len);
+            Some(decrement_cursor_x(ContentPosition { x, y }, content))
+        }
+    } else {
+        // this is the start of a selection
+        Some(ContentPosition { x: x.min(line_len), y })
+    }
+}
+
+fn decrement_cursor_x<T: Content>(cursor: ContentPosition, content: &T) -> ContentPosition {
+    if cursor.x > 0 {
+        ContentPosition {
+            x: cursor.x - 1,
+            y: cursor.y,
+        }
+    } else if cursor.y > 0 {
+        ContentPosition {
+            x: content.line_size(cursor.y - 1),
+            y: cursor.y - 1,
+        }
+    } else {
+        cursor
+    }
 }
 
 /// Widget that draws selection on the content.
 pub struct ContentSelectWidget<'a, T: Content> {
-    pub context: &'a SelectContext,
-    pub content: &'a T,
-    pub page_start: &'a PagePosition,
+    context: &'a SelectContext,
+    content: &'a T,
+    page_start: &'a ContentPosition,
+    color: Color,
 }
 
 impl<'a, T: Content> ContentSelectWidget<'a, T> {
     /// Creates new [`ContentSelectWidget`] instance.
-    pub fn new(context: &'a SelectContext, content: &'a T, page_start: &'a PagePosition) -> Self {
+    pub fn new(context: &'a SelectContext, content: &'a T, page_start: &'a ContentPosition, color: Color) -> Self {
         Self {
             context,
             content,
             page_start,
+            color,
         }
     }
 
@@ -184,12 +285,12 @@ impl<'a, T: Content> ContentSelectWidget<'a, T> {
     }
 }
 
-impl<'a, T: Content> Widget for ContentSelectWidget<'a, T> {
+impl<T: Content> Widget for ContentSelectWidget<'_, T> {
     fn render(self, area: Rect, buf: &mut ratatui::prelude::Buffer)
     where
         Self: Sized,
     {
-        let Some((start, end)) = self.context.get_selection() else {
+        let Some((start, end)) = self.context.get_selection().map(|s| s.sorted()) else {
             return;
         };
 
@@ -199,29 +300,29 @@ impl<'a, T: Content> Widget for ContentSelectWidget<'a, T> {
                 && y < area.bottom()
                 && let Some(max_x) = self.get_relative_max_len(area, current_line)
             {
-                let start = if start.y == current_line {
+                let start_x = if start.y == current_line {
                     // if this is the first line in the selection
                     self.get_relative_x(start.x, area).unwrap_or(0)
                 } else {
                     area.x
                 };
 
-                let end = if end.y == current_line {
+                let end_x = if end.y == current_line {
                     // if this is the last line in the selection
-                    self.get_relative_x(end.x, area).map(|x| x.min(max_x.saturating_sub(1)))
+                    self.get_relative_x(end.x, area).map(|x| x.min(max_x))
                 } else {
                     Some(max_x)
                 };
 
-                if start < area.right()
-                    && let Some(end) = end
+                if start_x < area.right()
+                    && let Some(end) = end_x
                     && end >= area.x
                 {
-                    let draw_from = start.max(area.x);
+                    let draw_from = start_x.max(area.x);
                     let draw_to = end.min(area.right());
 
                     for x in draw_from..=draw_to {
-                        buf[(x, y)].bg = ratatui::style::Color::Gray;
+                        buf[(x, y)].bg = self.color;
                     }
                 }
             }
@@ -229,12 +330,12 @@ impl<'a, T: Content> Widget for ContentSelectWidget<'a, T> {
     }
 }
 
-fn sort(p1: PagePosition, p2: PagePosition) -> (PagePosition, PagePosition) {
-    if (p1.y > p2.y) || (p1.y == p2.y && p1.x > p2.x) {
-        (p2, p1)
-    } else {
-        (p1, p2)
-    }
+fn is_sorted(p1: ContentPosition, p2: ContentPosition) -> bool {
+    p2.y > p1.y || (p2.y == p1.y && p2.x >= p1.x)
+}
+
+fn sort(p1: ContentPosition, p2: ContentPosition) -> (ContentPosition, ContentPosition) {
+    if is_sorted(p1, p2) { (p1, p2) } else { (p2, p1) }
 }
 
 fn is_allowed_key_code(key_code: KeyCode) -> bool {