+++ b4n/ui/presentation/mod.rs
@@ -1,4 +1,7 @@
-pub use content::{Content, ContentHeader, ContentViewer, MatchPosition, PagePosition, StyleFallback, StyledLine, StyledLineExt};
+pub use content::{
+    Content, ContentHeader, ContentPosition, ContentViewer, MatchPosition, Selection, StyleFallback, StyledLine, StyledLineExt,
+    VecStyledLineExt,
+};
 pub use list::{ListHeader, ListViewer};
 
 pub mod utils;