+++ b4n/ui/views/yaml/content.rs
@@ -6,7 +6,8 @@ use std::hash::{DefaultHasher, Hash, Hasher};
 use std::time::{Duration, Instant};
 use tokio::sync::{mpsc::UnboundedSender, oneshot::Receiver};
 
-use crate::ui::presentation::{Content, MatchPosition, PagePosition, StyleFallback, StyledLine, StyledLineExt};
+use crate::ui::presentation::utils::VecStringExt;
+use crate::ui::presentation::{Content, MatchPosition, Selection, StyleFallback, StyledLine, StyledLineExt, VecStyledLineExt};
 
 /// Number of lines before and after the modified section to include in the re-highlighting process.
 const HIGHLIGHT_CONTEXT_LINES_NO: usize = 200;
@@ -82,22 +83,17 @@ impl YamlContent {
         self.mark_line_as_modified(line_no);
     }
 
-    fn join_lines(&mut self, first: usize, second: usize) -> (usize, usize) {
-        let new_x = self.plain[first].chars().count();
+    fn join_lines(&mut self, line_no: usize) -> (usize, usize) {
+        let new_x = self.plain[line_no].chars().count();
 
-        self.styled[first].sl_push_str(&self.plain[second], &self.fallback);
-        self.styled.remove(second);
+        self.styled.join_lines(line_no);
+        self.plain.join_lines(line_no);
+        self.lowercase.join_lines(line_no);
 
-        let text = self.plain.remove(second);
-        self.plain[first].push_str(&text);
+        self.mark_line_as_modified(line_no);
+        self.mark_line_as_modified(line_no + 1);
 
-        let text = self.lowercase.remove(second);
-        self.lowercase[first].push_str(&text);
-
-        self.mark_line_as_modified(first);
-        self.mark_line_as_modified(second);
-
-        (new_x, first)
+        (new_x, line_no)
     }
 
     fn split_lines(&mut self, x: usize, y: usize) {
@@ -153,7 +149,7 @@ impl YamlContent {
     fn remove_char_internal(&mut self, x: usize, y: usize, is_backspace: bool, track_undo: bool) -> Option<(usize, usize)> {
         if is_backspace && x == 0 {
             if y > 0 && y < self.plain.len() {
-                let (x, y) = self.join_lines(y - 1, y);
+                let (x, y) = self.join_lines(y - 1);
                 return Some(self.track_remove(x, y, '\n', track_undo));
             }
 
@@ -170,7 +166,7 @@ impl YamlContent {
                 let ch = self.remove_ch(r.x.index, y);
                 Some(self.track_remove(r.x.char, y, ch, track_undo))
             } else if y + 1 < self.plain.len() {
-                let (x, y) = self.join_lines(y, y + 1);
+                let (x, y) = self.join_lines(y);
                 Some(self.track_remove(x, y, '\n', track_undo))
             } else {
                 None
@@ -196,6 +192,14 @@ impl YamlContent {
 
         (x, y)
     }
+
+    fn remove_text_internal(&mut self, range: Selection) -> Vec<String> {
+        self.styled.remove_text(range.clone());
+        self.plain.remove_text(range.clone());
+        self.lowercase.remove_text(range);
+
+        Vec::default()
+    }
 }
 
 impl Content for YamlContent {
@@ -219,27 +223,36 @@ impl Content for YamlContent {
         hasher.finish()
     }
 
-    fn to_plain_text(&self, range: Option<(PagePosition, PagePosition)>) -> String {
-        match range {
+    fn to_plain_text(&self, range: Option<Selection>) -> String {
+        match range.map(|r| r.sorted()) {
             None => self.plain.join("\n"),
             Some((start, end)) => {
                 let start_line = start.y.min(self.plain.len().saturating_sub(1));
                 let end_line = end.y.min(self.plain.len().saturating_sub(1));
 
-                let mut result = Vec::new();
-                for (i, line) in self.plain.iter().enumerate().take(end_line + 1).skip(start_line) {
+                let mut result = String::new();
+                for i in start_line..=end_line {
+                    let line = &self.plain[i];
                     if i == start_line && i == end_line {
-                        result.push(substring(line, start.x, (end.x + 1).saturating_sub(start.x)));
+                        result.push_str(substring(line, start.x, (end.x + 1).saturating_sub(start.x)));
+                        if line.chars().count() < end.x + 1 {
+                            result.push('\n');
+                        }
                     } else if i == start_line {
-                        result.push(slice_from(line, start.x));
+                        result.push_str(slice_from(line, start.x));
+                        result.push('\n');
                     } else if i == end_line {
-                        result.push(slice_to(line, end.x + 1));
+                        result.push_str(slice_to(line, end.x + 1));
+                        if line.chars().count() < end.x + 1 {
+                            result.push('\n');
+                        }
                     } else {
-                        result.push(line);
+                        result.push_str(line);
+                        result.push('\n');
                     }
                 }
 
-                result.join("\n")
+                result
             },
         }
     }
@@ -291,6 +304,12 @@ impl Content for YamlContent {
     fn remove_char(&mut self, x: usize, y: usize, is_backspace: bool) -> Option<(usize, usize)> {
         self.redo.clear();
         self.remove_char_internal(x, y, is_backspace, true)
+    }
+
+    fn remove_text(&mut self, range: Selection) -> Vec<String> {
+        self.mark_line_as_modified(range.start.y);
+        self.mark_line_as_modified(range.end.y);
+        self.remove_text_internal(range)
     }
 
     fn undo(&mut self) -> Option<(usize, usize)> {